+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/document/references/GenericZoomProvider.java
@@ -11,7 +11,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc., *
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA. *
  *****************************************************************************/
-package org.adempiere.model;
+package de.metas.document.references;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
