+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/document/references/POZoomSource.java
@@ -1,4 +1,4 @@
-package org.adempiere.model;
+package de.metas.document.references;
 
 import java.util.ArrayList;
 import java.util.Properties;
