+++ backend/de.metas.adempiere.adempiere/client/src/main/java-legacy/org/compiere/apps/AEnv.java
@@ -50,7 +50,6 @@ import org.adempiere.ad.element.api.AdWindowId;
 import org.adempiere.ad.table.api.IADTableDAO;
 import org.adempiere.ad.trx.api.ITrx;
 import org.adempiere.images.Images;
-import org.adempiere.model.RecordZoomWindowFinder;
 import org.adempiere.service.ClientId;
 import org.adempiere.util.lang.impl.TableRecordReference;
 import org.compiere.apps.form.FormFrame;
@@ -74,6 +73,7 @@ import de.metas.acct.api.IPostingRequestBuilder.PostImmediate;
 import de.metas.acct.api.IPostingService;
 import de.metas.adempiere.form.IClientUIInvoker.OnFail;
 import de.metas.cache.CCache;
+import de.metas.document.references.RecordZoomWindowFinder;
 import de.metas.i18n.IMsgBL;
 import de.metas.logging.LogManager;
 import de.metas.security.IUserRolePermissions;
