+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/relation/IRelationTypeDAO.java
@@ -31,7 +31,6 @@ import de.metas.util.ISingletonService;
 
 public interface IRelationTypeDAO extends ISingletonService
 {
-
 	I_AD_RelationType retrieveForInternalName(Properties ctx, String internalName);
 
 	default I_AD_RelationType retrieveForInternalName(final String internalName)
