+++ backend/metasfresh-webui-api/src/main/java/de/metas/ui/web/board/BoardDescriptorRepository.java
@@ -27,7 +27,6 @@ import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.exceptions.DBException;
 import org.adempiere.exceptions.DBUniqueConstraintException;
 import org.adempiere.model.InterfaceWrapperHelper;
-import org.adempiere.model.RecordZoomWindowFinder;
 import org.compiere.model.I_AD_User;
 import org.compiere.util.DB;
 import org.compiere.util.DisplayType;
@@ -45,6 +44,7 @@ import de.metas.cache.CCache;
 import de.metas.currency.Amount;
 import de.metas.currency.CurrencyCode;
 import de.metas.currency.CurrencyRepository;
+import de.metas.document.references.RecordZoomWindowFinder;
 import de.metas.i18n.IModelTranslationMap;
 import de.metas.i18n.IMsgBL;
 import de.metas.i18n.ITranslatableString;
