+++ backend/metasfresh-webui-api/src/main/java/de/metas/ui/web/window/model/DocumentReferencesService.java
@@ -9,9 +9,6 @@ import org.adempiere.ad.element.api.AdWindowId;
 import org.adempiere.ad.table.api.IADTableDAO;
 import org.adempiere.ad.trx.api.ITrx;
 import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.model.IZoomSource;
-import org.adempiere.model.ZoomInfo;
-import org.adempiere.model.ZoomInfoFactory;
 import org.compiere.model.I_AD_Column;
 import org.compiere.util.Evaluatee;
 import org.slf4j.Logger;
@@ -20,6 +17,9 @@ import org.springframework.stereotype.Service;
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Stopwatch;
 
+import de.metas.document.references.IZoomSource;
+import de.metas.document.references.ZoomInfo;
+import de.metas.document.references.ZoomInfoFactory;
 import de.metas.i18n.ITranslatableString;
 import de.metas.i18n.TranslatableStrings;
 import de.metas.logging.LogManager;
